Add files via upload
These files contain the SQL code for the database that I have fabricated as well as multiple queries I was required to make on said database as part of the final project. This also includes a write-up that explains the potential use for the database, as well as explanations for all of the entities within the database and their relationships.